#3837: moved erroneous mappers on erroneous subpackage
