+++ integrationtest/src/test/resources/fullFeatureTest/pom.xml
@@ -47,7 +47,6 @@
                         <exclude>**/*Test.java</exclude>
                         <exclude>**/testutil/**/*.java</exclude>
                         <exclude>**/spi/**/*.java</exclude>
-                        <exclude>**/ignoreunmapped/**/UserMapper*.java</exclude>
                         <exclude>${additionalExclude0}</exclude>
                         <exclude>${additionalExclude1}</exclude>
                         <exclude>${additionalExclude2}</exclude>
