+++ processor/src/test/java/org/mapstruct/ap/test/ignoreunmapped/IgnoredMappedPropertyTest.java
@@ -6,16 +6,16 @@
 package org.mapstruct.ap.test.ignoreunmapped;
 
 import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapper;
-import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithMultiMapping;
+import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithWarnSourcePolicy;
 import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithWarnPolicyInMapperConfig;
-import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithErrorPolicyInMapperConfig;
+import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithWarnSourcePolicyInMapper;
 import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithoutBeanMapping;
 import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithIgnorePolicyInMapperConfig;
-import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithWarnSourcePolicy;
 import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithIgnoreSourcePolicy;
-import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithErrorSourcePolicy;
-import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithErrorSourcePolicyInMapper;
-import org.mapstruct.ap.test.ignoreunmapped.mapper.UserMapperWithWarnSourcePolicyInMapper;
+import org.mapstruct.ap.test.ignoreunmapped.mapper.erroneous.UserMapperWithErrorPolicyInMapperConfig;
+import org.mapstruct.ap.test.ignoreunmapped.mapper.erroneous.UserMapperWithErrorSourcePolicy;
+import org.mapstruct.ap.test.ignoreunmapped.mapper.erroneous.UserMapperWithErrorSourcePolicyInMapper;
+import org.mapstruct.ap.test.ignoreunmapped.mapper.erroneous.UserMapperWithMultiMapping;
 import org.mapstruct.ap.testutil.IssueKey;
 import org.mapstruct.ap.testutil.ProcessorTest;
 import org.mapstruct.ap.testutil.WithClasses;
@@ -60,19 +60,22 @@ public class IgnoredMappedPropertyTest {
                             type = UserMapperWithErrorSourcePolicy.class,
                             kind = Kind.ERROR,
                             line = 20,
-                            message = "Source property \"email\" is mapped despite being listed in ignoreUnmappedSourceProperties."
+                            message = "Source property \"email\" is mapped despite being " +
+                                    "listed in ignoreUnmappedSourceProperties."
                     ),
                     @Diagnostic(
                             type = UserMapperWithErrorSourcePolicyInMapper.class,
                             kind = Kind.ERROR,
                             line = 19,
-                            message = "Source property \"email\" is mapped despite being listed in ignoreUnmappedSourceProperties."
+                            message = "Source property \"email\" is mapped despite being " +
+                                    "listed in ignoreUnmappedSourceProperties."
                     ),
                     @Diagnostic(
                             type = UserMapperWithErrorPolicyInMapperConfig.class,
                             kind = Kind.ERROR,
                             line = 23,
-                            message = "Source property \"email\" is mapped despite being listed in ignoreUnmappedSourceProperties."
+                            message = "Source property \"email\" is mapped despite being " +
+                                    "listed in ignoreUnmappedSourceProperties."
                     )
             }
     )
@@ -92,19 +95,22 @@ public class IgnoredMappedPropertyTest {
                             type = UserMapperWithWarnSourcePolicy.class,
                             kind = Kind.WARNING,
                             line = 20,
-                            message = "Source property \"email\" is mapped despite being listed in ignoreUnmappedSourceProperties."
+                            message = "Source property \"email\" is mapped despite being " +
+                                    "listed in ignoreUnmappedSourceProperties."
                     ),
                     @Diagnostic(
                             type = UserMapperWithWarnSourcePolicyInMapper.class,
                             kind = Kind.WARNING,
                             line = 19,
-                            message = "Source property \"email\" is mapped despite being listed in ignoreUnmappedSourceProperties."
+                            message = "Source property \"email\" is mapped despite being " +
+                                    "listed in ignoreUnmappedSourceProperties."
                     ),
                     @Diagnostic(
                             type = UserMapperWithWarnPolicyInMapperConfig.class,
                             kind = Kind.WARNING,
                             line = 23,
-                            message = "Source property \"email\" is mapped despite being listed in ignoreUnmappedSourceProperties."
+                            message = "Source property \"email\" is mapped despite being " +
+                                    "listed in ignoreUnmappedSourceProperties."
                     )
             }
     )
@@ -120,7 +126,8 @@ public class IgnoredMappedPropertyTest {
                             type = UserMapperWithMultiMapping.class,
                             kind = Kind.ERROR,
                             line = 21,
-                            message = "Source properties \"email, username\" are mapped despite being listed in ignoreUnmappedSourceProperties."
+                            message = "Source properties \"email, username\" are mapped despite " +
+                                    "being listed in ignoreUnmappedSourceProperties."
                     )
             }
     )
