+++ processor/src/test/java/org/mapstruct/ap/test/ignoreunmapped/mapper/erroneous/UserMapperWithErrorSourcePolicy.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.ignoreunmapped.mapper;
+package org.mapstruct.ap.test.ignoreunmapped.mapper.erroneous;
 
 import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
@@ -12,9 +12,10 @@ import org.mapstruct.ReportingPolicy;
 import org.mapstruct.ap.test.ignoreunmapped.UserDto;
 import org.mapstruct.ap.test.ignoreunmapped.UserEntity;
 
-@Mapper(unmappedSourcePolicy = ReportingPolicy.ERROR)
-public interface UserMapperWithErrorSourcePolicyInMapper {
-    @BeanMapping(ignoreUnmappedSourceProperties = {"password", "email"})
+@Mapper
+public interface UserMapperWithErrorSourcePolicy {
+
+    @BeanMapping(ignoreUnmappedSourceProperties = {"password", "email"}, unmappedSourcePolicy = ReportingPolicy.ERROR)
     @Mapping(source = "email", target = "email")
     UserDto map(UserEntity user);
 }
