+++ processor/src/test/java/org/mapstruct/ap/test/ignoreunmapped/mapper/erroneous/UserMapperWithErrorSourcePolicyInMapper.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.ignoreunmapped.mapper;
+package org.mapstruct.ap.test.ignoreunmapped.mapper.erroneous;
 
 import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
@@ -12,11 +12,9 @@ import org.mapstruct.ReportingPolicy;
 import org.mapstruct.ap.test.ignoreunmapped.UserDto;
 import org.mapstruct.ap.test.ignoreunmapped.UserEntity;
 
-@Mapper
-public interface UserMapperWithMultiMapping {
-    @BeanMapping(ignoreUnmappedSourceProperties = {"password", "email", "username"},
-            unmappedSourcePolicy = ReportingPolicy.ERROR)
+@Mapper(unmappedSourcePolicy = ReportingPolicy.ERROR)
+public interface UserMapperWithErrorSourcePolicyInMapper {
+    @BeanMapping(ignoreUnmappedSourceProperties = {"password", "email"})
     @Mapping(source = "email", target = "email")
-    @Mapping(source = "username", target = "username")
     UserDto map(UserEntity user);
 }
